update install locations, add --root, no longer use nodepy.Directories

+++ bootstrap.py
@@ -38,9 +38,10 @@ import sys
 @click.option('--install', is_flag=True, help="Install PPYM after bootstrapping "
     "and clean up the bootstrapped nodepy_modules/ directory.")
 @click.option('-g', '--global', 'global_', is_flag=True, help="Install PPYM globally.")
+@click.option('--root', is_flag=True, help="Install PPYM in the Python root folder.")
 @click.option('-U', '--upgrade', is_flag=True, help="Uninstall previous versions instead of skipping the new version.")
 @click.option('--develop', is_flag=True, help="If --install, install in development mode.")
-def main(bootstrap, install, global_, upgrade, develop):
+def main(bootstrap, install, global_, root, upgrade, develop):
   """
   Bootstrap the PPYM installation.
   """
@@ -72,12 +73,14 @@ def main(bootstrap, install, global_, upgrade, develop):
       cmd.append('--upgrade')
     if global_:
       cmd.append('--global')
+    if root:
+      cmd.append('--root')
     if develop:
       cmd.append('--develop')
     cmd.append(__directory__)
     require('./index').main(cmd, standalone_mode=False)
 
-    if not existed_before and bootstrap:
+    if not existed_before and bootstrap and os.path.isdir('nodepy_modules'):
       print('Cleaning up bootstrap modules directory ...')
       shutil.rmtree('nodepy_modules')
